Refactors the encoding and decoding functionality into their own dedicated files
